the major functionality is all working.  still needs some polish
the default SMPL version is now SUPR instad of SMPLH
changed the 'SMPL version' meta data field to `SMPL_version` because maya can't handle underscores there
changed all the handpose variable names to hand_pose in code for consistancy sake
there was redundant hand pose functionality so now it all calls the same function
the handposes.npz file was poorly constructed so Nikolas made new ones for the relaxed pose, and we juse use np.zeroes(45) for the flat hand pose
made a more verbose error message if the incoming .npz file is missing any necissary data
whenever you load a pose or an avatar, you can select certain things like the SMPL version and the hand pose, if they choose one of those things, the UI gets updated to match that
hard coded disabled shape keys coming in for SMPLH because we don't have a joint regressor for it
made it so that it doesn't have to override the hand poses when you load a pose or avatar
fixed a bug where the translation vector was 100 times too small for SUPR and SMPLH
made OP_UpdateJointLocations exit early if the model is SMPLH because we don't have the joint regressor for it
removed the entire export to unity functionality because I changed a lot and I don't have time to test it.  I can add it back later if needed.
removed a lot of places where the names were specific to smplx.  I changed them to just be smpl instead
I read in the hand pose files once on load and save them.  it was reading it in every time you called the function